third_party/WebKit/LayoutTests/editing/execCommand/format-block-contenteditable-false.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
  [
    '<div contenteditable>',
    '   ^Will select from here',
    '   <div contenteditable="false">',
    '       over this contenteditable=false div',
    '   </div>',
    '   until here.|',
    '</div>',
  ].join('\n'),
  selection => {
    selection.document.execCommand('italic');
    selection.document.execCommand('formatblock', false, '<h1>');
  },
  [
    '<div contenteditable><h1><i>',
    '   ^Will select from here',
    '   </i><br><i>',
    '   until here.|',
    '</i></h1><div contenteditable="false">',
    '       over this contenteditable=false div',
    '   </div></div>',
  ].join('\n')),
  'formatBlock with uneditable');
</script>
