Safari-609.1.11.1/LayoutTests/editing/execCommand/format-block-contenteditable-false-expected.txt

formatblock on selection with a contenteditable="false" child.
| <h1>
|   <i>
|     "
        <#selection-anchor>Will select from here
        "
|   <br>
|   <i>
|     "
        until here.<#selection-focus>
    "
| <div>
|   contenteditable="false"
|   "
            over this contenteditable=false div
        "
